+++ src/components/Logo.tsx
@@ -1,6 +1,6 @@
-import React from "react";
+import React from 'react';
 
-const GitcoinLogo = () => (
+export const GitcoinLogo = () => (
   <svg
     width="24"
     height="24"
@@ -14,13 +14,13 @@ const GitcoinLogo = () => (
         id="Vector"
         d="M17.8465 20.0901C17.773 20.1124 17.7366 20.133 17.701 20.133C15.6958 20.133 13.6835 20.2371 11.6878 20.1036C8.55852 19.8947 6.07317 17.6605 5.27267 14.7843C4.33533 11.4171 5.87384 7.88699 8.97855 6.29078C9.22455 6.16445 9.32422 6.02223 9.3084 5.74733C9.28625 5.36675 9.2989 4.98379 9.30444 4.60162C9.3084 4.29096 9.36456 4.0089 9.76165 4.00016C10.0749 3.9938 10.2125 4.17575 10.2157 4.59367C10.2181 4.97187 10.2157 5.35006 10.2157 5.75368C10.6824 5.66072 11.1016 5.5773 11.5651 5.48513C11.5651 5.18162 11.5596 4.86778 11.5667 4.55474C11.5746 4.24805 11.6798 4.00016 12.0374 4.00334C12.3902 4.00652 12.5017 4.23534 12.4962 4.55633C12.4914 4.85507 12.4954 5.15461 12.4954 5.48513C13.776 5.51294 14.9491 5.82281 15.99 6.48703C16.6798 6.9272 17.2952 7.48575 18 8.03636C16.6664 9.28218 15.4063 10.4605 14.151 11.6332C14.5876 12.8568 14.5813 13.0268 14.0457 14.1193C14.6604 14.5698 15.2876 15.0163 15.9015 15.4827C16.5058 15.9419 17.0975 16.4178 17.6883 16.8937C17.7603 16.9517 17.8394 17.0574 17.8402 17.1416C17.8505 18.1261 17.8473 19.1097 17.8473 20.0909L17.8465 20.0901ZM16.0565 8.06258C15.8769 7.93148 15.7464 7.82263 15.6025 7.73364C14.0355 6.76909 12.368 6.51007 10.5867 7.00109C10.3106 7.07736 10.2109 7.19893 10.2212 7.49052C10.2434 8.14839 10.226 8.80785 10.2283 9.46651C10.2299 9.7875 10.1144 10.0163 9.76244 10.0171C9.41123 10.0171 9.291 9.77955 9.28546 9.46651C9.27913 9.11056 9.28388 8.75541 9.28388 8.39946C9.28388 8.14521 9.28388 7.89096 9.28388 7.57156C9.09878 7.68915 8.98567 7.75192 8.88283 7.82819C6.85785 9.33779 5.99486 11.3495 6.36901 13.8579C6.7479 16.3956 8.96905 18.5718 11.5161 18.78C13.1574 18.9134 14.8154 18.8388 16.4654 18.8538C16.4987 18.8538 16.5319 18.8197 16.5825 18.7911C16.5825 18.598 16.5612 18.3978 16.5873 18.2047C16.6403 17.8082 16.4828 17.5596 16.1601 17.3268C15.1919 16.6284 14.2427 15.9014 13.3022 15.1657C13.0427 14.9623 12.796 14.905 12.474 14.9392C11.7344 15.0171 11.1056 14.7628 10.6412 14.1685C10.0195 13.3724 9.96731 12.4778 10.4775 11.6729C11.0035 10.8426 11.8808 10.4994 12.8608 10.7179C12.989 10.7465 13.1835 10.7258 13.2729 10.6448C14.196 9.80577 15.1041 8.95007 16.0549 8.06337L16.0565 8.06258ZM12.2945 13.6568C12.7714 13.66 13.1464 13.2898 13.1487 12.8123C13.1519 12.3387 12.7762 11.9558 12.3063 11.9526C11.8365 11.9494 11.4536 12.3292 11.4512 12.8011C11.4489 13.2779 11.8198 13.6545 12.2945 13.6576V13.6568Z"
         fill="white"
-        fill-opacity="0.75"
+        fillOpacity="0.75"
       />
     </g>
   </svg>
 );
 
-const GivethLogo = () => (
+export const GivethLogo = () => (
   <svg
     width="24"
     height="24"
@@ -48,7 +48,7 @@ const GivethLogo = () => (
   </svg>
 );
 
-const OptimismLogo = () => (
+export const OptimismLogo = () => (
   <svg
     width="24"
     height="24"
@@ -56,7 +56,7 @@ const OptimismLogo = () => (
     fill="none"
     xmlns="http://www.w3.org/2000/svg"
   >
-    <g id="Network Logos" clip-path="url(#clip0_668_1248)">
+    <g id="Network Logos" clipPath="url(#clip0_668_1248)">
       <path
         id="Vector"
         d="M12 24C18.6274 24 24 18.6274 24 12C24 5.37258 18.6274 0 12 0C5.37258 0 0 5.37258 0 12C0 18.6274 5.37258 24 12 24Z"
@@ -81,7 +81,7 @@ const OptimismLogo = () => (
   </svg>
 );
 
-const EthLogo = () => (
+export const EthLogo = () => (
   <svg
     width="24"
     height="24"
@@ -90,11 +90,11 @@ const EthLogo = () => (
     xmlns="http://www.w3.org/2000/svg"
   >
     <circle cx="12" cy="12" r="12" fill="white" />
-    <g clip-path="url(#clip0_685_548)">
+    <g clipPath="url(#clip0_685_548)">
       <path
         d="M11.9983 3L11.8804 3.42374V15.7197L11.9983 15.8442L17.3967 12.4704L11.9983 3Z"
         fill="#5326EC"
-        fill-opacity="0.2"
+        fillOpacity="0.2"
       />
       <path
         d="M11.9985 3L6.6001 12.4704L11.9985 15.8442V9.87611V3Z"
@@ -103,7 +103,7 @@ const EthLogo = () => (
       <path
         d="M11.9986 16.9248L11.9321 17.0105V21.3906L11.9986 21.5958L17.4002 13.5527L11.9986 16.9248Z"
         fill="#5326EC"
-        fill-opacity="0.4"
+        fillOpacity="0.4"
       />
       <path
         d="M11.9985 21.5958V16.9248L6.6001 13.5527L11.9985 21.5958Z"
@@ -116,7 +116,7 @@ const EthLogo = () => (
       <path
         d="M6.6001 12.4712L11.9984 15.845V9.87695L6.6001 12.4712Z"
         fill="#5326EC"
-        fill-opacity="0.2"
+        fillOpacity="0.2"
       />
     </g>
     <defs>
@@ -132,20 +132,20 @@ const EthLogo = () => (
   </svg>
 );
 
-type Status = "GTCLOGO" | "GIVETHLOGO" | "OPLOGO" | "ETHLOGO";
+type Status = 'GTCLOGO' | 'GIVETHLOGO' | 'OPLOGO' | 'ETHLOGO';
 
 type CheckProps = {
   status: Status;
 };
 
 export function Check(props: CheckProps) {
-  if (props.status === "GTCLOGO") {
+  if (props.status === 'GTCLOGO') {
     return <GitcoinLogo />;
-  } else if (props.status === "GIVETHLOGO") {
+  } else if (props.status === 'GIVETHLOGO') {
     return <GivethLogo />;
-  } else if (props.status === "OPLOGO") {
+  } else if (props.status === 'OPLOGO') {
     return <OptimismLogo />;
-  } else if (props.status === "ETHLOGO") {
+  } else if (props.status === 'ETHLOGO') {
     return <EthLogo />;
   }
 }
